+++ stroom-query-api/src/test/java/stroom/query/api/v2/FieldBuilderTest.java
@@ -18,6 +18,7 @@ class FieldBuilderTest {
         final Boolean numberFormatUseSeperator = true;
 
         final Field field = new Field.Builder()
+                .id(name)
                 .name(name)
                 .expression(expression)
                 .sort(new Sort.Builder()