+++ stroom-query-common/src/test/java/stroom/query/common/v2/TestTablePayloadHandler.java
@@ -49,6 +49,7 @@ class TestTablePayloadHandler {
 
         final TableSettings tableSettings = new TableSettings.Builder()
                 .addFields(new Field.Builder()
+                        .id("Text")
                         .name("Text")
                         .expression(ParamUtil.makeParam("Text"))
                         .format(Format.Type.TEXT)
@@ -104,6 +105,7 @@ class TestTablePayloadHandler {
 
         final TableSettings tableSettings = new TableSettings.Builder()
                 .addFields(new Field.Builder()
+                        .id("Text")
                         .name("Text")
                         .expression(ParamUtil.makeParam("Text"))
                         .sort(sort)
@@ -148,6 +150,7 @@ class TestTablePayloadHandler {
 
         final TableSettings tableSettings = new TableSettings.Builder()
                 .addFields(new Field.Builder()
+                        .id("Number")
                         .name("Number")
                         .expression(ParamUtil.makeParam("Number"))
                         .sort(sort)
@@ -185,6 +188,7 @@ class TestTablePayloadHandler {
 //        final DataSourceFieldsMap dataSourceFieldsMap = new DataSourceFieldsMap();
 
         final Field count = new Field.Builder()
+                .id("Count")
                 .name("Count")
                 .expression("count()")
                 .sort(sort)
@@ -197,6 +201,7 @@ class TestTablePayloadHandler {
 
         final TableSettings tableSettings = new TableSettings.Builder()
                 .addFields(new Field.Builder()
+                        .id("Text")
                         .name("Text")
                         .expression(ParamUtil.makeParam("Text"))
                         .group(0)
@@ -241,10 +246,12 @@ class TestTablePayloadHandler {
 
         final TableSettings tableSettings = new TableSettings.Builder()
                 .addFields(new Field.Builder()
+                        .id("Count")
                         .name("Count")
                         .expression("count()")
                         .build())
                 .addFields(new Field.Builder()
+                        .id("Text")
                         .name("Text")
                         .expression(ParamUtil.makeParam("Text"))
                         .sort(sort)
@@ -289,10 +296,12 @@ class TestTablePayloadHandler {
 
         final TableSettings tableSettings = new TableSettings.Builder()
                 .addFields(new Field.Builder()
+                        .id("Count")
                         .name("Count")
                         .expression("count()")
                         .build())
                 .addFields(new Field.Builder()
+                        .id("Text")
                         .name("Text")
                         .expression(ParamUtil.makeParam("Text"))
                         .sort(sort)