+++ stroom-query-testing/src/test/java/stroom/query/testing/memory/TestQueryResourceIT.java
@@ -64,6 +64,7 @@ class TestQueryResourceIT extends QueryResourceIT<TestDocRefEntity, Config> {
                                 .extractValues(false)
                                 .showDetail(false)
                                 .addFields(new Field.Builder()
+                                        .id(TestDocRefEntity.INDEX_NAME)
                                         .name(TestDocRefEntity.INDEX_NAME)
                                         .expression("${" + TestDocRefEntity.INDEX_NAME + "}")
                                         .build())