+++ stroom-query-testing/src/test/java/stroom/query/testing/memory/TestQueryResourceNoAuthIT.java
@@ -53,6 +53,7 @@ class TestQueryResourceNoAuthIT extends QueryResourceNoAuthIT<TestDocRefEntity, 
                                 .extractValues(false)
                                 .showDetail(false)
                                 .addFields(new Field.Builder()
+                                        .id(TestDocRefEntity.INDEX_NAME)
                                         .name(TestDocRefEntity.INDEX_NAME)
                                         .expression("${" + TestDocRefEntity.INDEX_NAME + "}")
                                         .build())